requirements.txt
absl-py==1.4.0
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.2
contourpy==1.0.7
crispy-bootstrap4==2022.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
django-address==0.2.8
django-crispy-forms==2.0
django-phonenumber-field==7.0.2
executing==1.2.0
flatbuffers==1.12
fonttools==4.39.2
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.3.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
openvino-tensorflow==2.3.0
opt-einsum==3.3.0
packaging==23.0
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.7
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==25.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
supervisor==4.2.5
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
django-address